Fix the MSVC OpenMP build of omc_matrad
MSVC only implements OpenMP 2.0, which for C compilation units does not
accept declaring the loop variable inside the for statement of a
parallel for (error C3015); the accumulateResults() loop now declares
it beforehand like the history loops already do. Also makes the
double-to-int truncation in mscat() explicit, which was the one C4244
warning MSVC raised in changed code.

Verified with a local MSVC Release build: all four targets compile and
the smoke case runs -- with dose output byte-identical to the MinGW
build, and about twice as fast (interleaved: ~3.9 s vs ~8.8 s with
gcc 13), so MSVC is the better choice for Windows binaries.

+++ src/ompmc.c
@@ -3836,7 +3836,7 @@ void mscat(int imed, int qel, int *spin_index, int *find_index,
 			counter2++;
 			xi = setRandom();
 			ak = xi*MXU_MS;
-			k = ak;
+			k = (int)ak;
 			ak -= k;
             
 			int msbase = m_scat->i*(MXQ_MS + 1)*(MXU_MS + 1) +
